functions/requirements.txt
firebase_functions~=0.1.0
Pillow==10.2.0
tensorflow==2.16.1